refactor(render): 重构渲染边界并收口 legacy 兼容层
- 新增 WindowContext，将窗口生命周期、事件轮询和 drawable size 查询从 RenderContext 中拆出
- 收缩 RenderContext 为纯 GPU / surface 上下文，拆分 acquire、encoder、submit、present 低层接口
- 新增 SurfaceFrame 和 LegacyFrameContext，分别服务新渲染路径与 legacy 2D 兼容路径
- 将 GuiRenderer 重命名为 LegacyGuiRenderer，统一场景和应用层接口命名
- 删除 PipelineLibrary，拆分为 Scene3DPipelineFactory 和 LegacyPipeline2D
- 调整 Renderer、ForwardPass、WireframePass、GuiPass 的帧编排与依赖边界
- 更新 Scene2D 和 Scene3DLegacy 对新渲染边界的接入方式
- 修复 Scene2D index buffer 上传未按 COPY_BUFFER_ALIGNMENT 对齐的校验错误

+++ src/render/gpu/GpuResourceCache.cpp
@@ -1,13 +1,12 @@
 #include "GpuResourceCache.h"
 
 #include <algorithm>
-#include <cstdint>
 #include <vector>
 
 #include "render/RenderContext.h"
 
 std::size_t GpuResourceCache::CacheKeyHash::operator()(const CacheKey& key) const {
-    const auto meshBits = static_cast<std::size_t>(reinterpret_cast<std::uintptr_t>(key.mesh));
+    const auto meshBits = reinterpret_cast<std::uintptr_t>(key.mesh);
     const auto modeBits = static_cast<std::size_t>(key.renderMode == Object3D::ERenderMode::Wireframe ? 1U : 0U);
     return meshBits ^ (modeBits << 1U);
 }
